dev 合入的 issue #186（一系统同步管线 course-pk-sync）与 PR #197 各自独立设计了
pk 数据域模型，本次按「对齐 dev 模型 + 适配 PR #197 查询侧」整合：

- 实体定义统一采用 dev 版本：CalendarId uint64、可空指针字段、gorm 时间戳
  替代 SchemaVersion/SyncedAt
- CourseNature 拆分全局字典 pk_course_nature + 学期快照 pk_course_nature_by_calendar，
  查询侧 JOIN 改用 by_calendar 表
- FetchLog 采用 dev 同步游标语义；latest-update 改读 completed 的 finished_at
- 排课文本解析收敛到 arrange.go（指针版 ArrangementInfo），parse.go 仅保留同步管线
  独用函数；dev timeslots.go 适配指针解引用
- 保留 PR #197 独有 SettingEntity(pk_setting) 与 timeslots 懒构建
- 补齐 builder、去重 ListTeachersByClassIds、适配契约测试 fixture

验证：go build ./...、go test ./...、redocly lint/bundle、generate:ts（无 diff）、
check:gen 全部通过。

+++ apps/gooseforum/app/service/pkservice/catalog.go
@@ -6,7 +6,7 @@ import (
 
 // CalendarItem P1 calendars 输出项。
 type CalendarItem struct {
-	CalendarId   int    `json:"calendarId"`
+	CalendarId   uint64 `json:"calendarId"`
 	CalendarName string `json:"calendarName"`
 }
